Django D3/requirements.txt
APScheduler        3.10.4
asgiref            3.8.1
backports.zoneinfo 0.2.1
certifi            2024.2.2
cffi               1.16.0
charset-normalizer 3.3.2
cryptography       42.0.6
defusedxml         0.7.1
Django             4.2.11
django-allauth     0.54.0
django-apscheduler 0.6.2
django-filter      21.1
idna               3.7
oauthlib           3.2.2
pip                24.0
pycparser          2.22
PyJWT              2.8.0
python3-openid     3.2.0
pytz               2024.1
requests           2.31.0
requests-oauthlib  2.0.0
setuptools         47.1.0
six                1.16.0
sqlparse           0.4.4
typing_extensions  4.10.0
tzdata             2024.1
tzlocal            5.2
urllib3            2.2.1

# pip install -r requirements.txt - to install necessary packages